FIX: step_counter is 32 bit ADD: step counter for watch face 1 is in nerd steps.

+++ include/sw_screens.h
@@ -11,6 +11,6 @@
 #include "sw_rtc.h"
 
 void sw_splashscreen(display_t *dsp, now_t *n);
-void sw_watch1_init(display_t *dsp, now_t *now, uint16_t steps);
+void sw_watch1_init(display_t *dsp, now_t *now, uint32_t steps);
 void sw_watch1();
 void sw_watch1_ble_status_cb();
